requirements.txt
streamlit>=1.28.0
pillow>=9.0.0
beautifulsoup4>=4.11.0
requests>=2.28.0
scikit-learn>=1.2.0
google-generativeai>=0.3.0
numpy>=1.24.0